# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Finances API",
    "version": "1.0.0",
    "description": "API para gerenciamento financeiro pessoal, com controle de contas, transações e categorias."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Servidor de Desenvolvimento"
    }
  ],
  "paths": {
...
    "/transactions": {
      "post": {
        "summary": "Cria uma nova transação",
        "tags": [
          "Transactions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "account_id",
                  "name",
                  "amount",
                  "type",
                  "transaction_date"
                ],
                "properties": {
                  "account_id": {
                    "type": "string",
                    "description": "ID da conta"
                  },
                  "category_id": {
                    "type": "string",
                    "description": "ID da categoria (opcional)"
                  },
                  "category_name": {
                    "type": "string",
                    "description": "Nome da categoria (criada automaticamente se não existir)"
                  },
                  "payment_method_id": {
                    "type": "string",
                    "description": "ID do método de pagamento (opcional)"
                  },
                  "name": {
                    "type": "string",
                    "description": "Nome/descrição da transação"
                  },
                  "amount": {
                    "type": "number",
                    "description": "Valor da transação"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "income",
                      "expense"
                    ],
                    "description": "Tipo da transação"
                  },
                  "transaction_date": {
                    "type": "string",
                    "format": "date",
                    "description": "Data da transação"
                  },
                  "description": {
                    "type": "string",
                    "description": "Descrição adicional (opcional)"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Tags da transação (opcional)"
                  }
                },
                "example": {
                  "account_id": "abc123-def456-ghi789",
                  "category_name": "Alimentação",
                  "payment_method_id": "payment123",
                  "name": "Compra no Supermercado",
                  "amount": 150.5,
                  "type": "expense",
                  "transaction_date": "2023-01-15",
                  "description": "Compras da semana",
                  "tags": [
                    "alimentação",
                    "supermercado"
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transação criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "400": {
            "description": "Dados de entrada inválidos"
          },
          "404": {
            "description": "Conta não encontrada"
          }
        }
      },
      "get": {
        "summary": "Lista transações com filtros opcionais",
        "tags": [
          "Transactions"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "account_id",
            "schema": {
              "type": "string"
            },
            "description": "Filtrar por ID da conta"
          },
          {
            "in": "query",
            "name": "category_id",
            "schema": {
              "type": "string"
            },
            "description": "Filtrar por ID da categoria"
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "income",
                "expense"
              ]
            },
            "description": "Filtrar por tipo de transação"
          },
          {
            "in": "query",
            "name": "start_date",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Data inicial para filtro"
          },
          {
            "in": "query",
            "name": "end_date",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Data final para filtro"
          },
          {
            "in": "query",
            "name": "min_amount",
            "schema": {
              "type": "number"
            },
            "description": "Valor mínimo para filtro"
          },
          {
            "in": "query",
            "name": "max_amount",
            "schema": {
              "type": "number"
            },
            "description": "Valor máximo para filtro"
          },
          {
            "in": "query",
            "name": "payment_method_id",
            "schema": {
              "type": "string"
            },
            "description": "Filtrar por método de pagamento"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Buscar por nome ou descrição"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de transações",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/transactions/{id}": {
      "get": {
        "summary": "Busca uma transação pelo ID",
        "tags": [
          "Transactions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "ID da transação"
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhes da transação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "404": {
            "description": "Transação não encontrada"
          }
        }
      },
      "put": {
        "summary": "Atualiza uma transação existente",
        "tags": [
          "Transactions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "ID da transação a ser atualizada"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "category_id": {
                    "type": "string"
                  },
                  "category_name": {
                    "type": "string"
                  },
                  "payment_method_id": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "number"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "income",
                      "expense"
                    ]
                  },
                  "transaction_date": {
                    "type": "string",
                    "format": "date"
                  },
                  "description": {
                    "type": "string"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transação atualizada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "400": {
            "description": "Dados de entrada inválidos"
          },
          "404": {
            "description": "Transação não encontrada"
          }
        }
      },
      "delete": {
        "summary": "Remove uma transação",
        "tags": [
          "Transactions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "ID da transação a ser removida"
          }
        ],
        "responses": {
          "204": {
            "description": "Transação removida com sucesso"
          },
          "404": {
            "description": "Transação não encontrada"
          }
        }
      }
    },
    "/transactions/account/{accountId}": {
      "get": {
        "summary": "Lista transações de uma conta específica",
        "tags": [
          "Transactions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "accountId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "ID da conta"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de transações da conta",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Conta não encontrada"
          }
        }
      }
    },
    "/transactions/user/{userId}": {
      "get": {
        "summary": "Lista transações de um usuário específico",
        "tags": [
          "Transactions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "ID do usuário"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de transações do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/transactions/account/{accountId}/balance": {
      "get": {
        "summary": "Obtém o saldo atual de uma conta",
        "tags": [
          "Transactions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "accountId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "ID da conta"
          }
        ],
        "responses": {
          "200": {
            "description": "Saldo da conta",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "balance": {
                      "type": "number",
                      "description": "Saldo atual da conta"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Conta não encontrada"
          }
        }
      }
    },
    "/transactions/account/{accountId}/balance-details": {
      "get": {
        "summary": "Obtém detalhes do saldo de uma conta",
        "tags": [
          "Transactions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "accountId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "ID da conta"
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhes do saldo da conta",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "initial_balance": {
                      "type": "number",
                      "description": "Saldo inicial da conta"
                    },
                    "total_income": {
                      "type": "number",
                      "description": "Total de receitas"
                    },
                    "total_expense": {
                      "type": "number",
                      "description": "Total de despesas"
                    },
                    "current_balance": {
                      "type": "number",
                      "description": "Saldo atual"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Conta não encontrada"
          }
        }
      }
    },
    "/transactions/account/{accountId}/current-month": {
      "get": {
        "summary": "Lista transações do mês atual de uma conta",
        "tags": [
          "Transactions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "accountId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "ID da conta"
          }
        ],
        "responses": {
          "200": {
            "description": "Transações do mês atual",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Conta não encontrada"
          }
        }
      }
    },
    "/transactions/user/{userId}/current-month": {
      "get": {
        "summary": "Lista transações do mês atual de um usuário",
        "tags": [
          "Transactions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "ID do usuário"
          }
        ],
        "responses": {
          "200": {
            "description": "Transações do mês atual do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}